Switched the ApplicationController base.
