+++ app/controllers/factory_bot/instrumentation/root_controller.rb
@@ -4,12 +4,6 @@ module FactoryBot
   module Instrumentation
     # The Instrumentation engine controller with frontend and API actions.
     class RootController < ApplicationController
-      # This is required to make API controllers template renderable
-      include ActionView::Layouts
-
-      # Configure the default application layout
-      layout 'factory_bot/instrumentation/application'
-
       # Show the instrumentation frontend which features the output of
       # configured dynamic seeds scenarios. The frontend allows humans to
       # generate new seed data on the fly.
@@ -53,8 +47,6 @@ module FactoryBot
         FactoryBot::Instrumentation.configuration.render_error.call(self, e)
       end
 
-      protected
-
       # Parse the given parameters from the request and build
       # a valid FactoryBot options set.
       #
@@ -79,53 +71,6 @@ module FactoryBot
         ]
       end
       # rubocop:enable Metrics/MethodLength
-
-      # Unfortunately +Rails.configuration.instrumentation+ is only read once
-      # and do not hot-reload on changes. Thats why we read this file manually
-      # to get always a fresh state.
-      #
-      # @return [Hash{String => Mixed}] the instrumentation scenarios
-      def instrumentation
-        config_file = FactoryBot::Instrumentation.configuration.config_file.to_s
-        template = ERB.new(Pathname.new(config_file).read)
-        load_method = YAML.respond_to?(:unsafe_load) ? :unsafe_load : :load
-        YAML.send(load_method, template.result(binding))[Rails.env] || {}
-      end
-
-      # Map all the instrumentation scenarios into groups and pass them back.
-      #
-      # @return [Hash{String => Array}] the grouped scenarios
-      def scenarios
-        res = instrumentation['scenarios'] || []
-        res.each_with_object({}) do |scenario, memo|
-          group = scenario_group(scenario['name'])
-          scenario['group'] = group
-          memo[group] = [] unless memo.key? group
-          memo[group] << scenario
-        end
-      end
-
-      # Map all the configured scenario groups to a useable hash.
-      #
-      # @return [Hash{Regexp => String}] the group mapping
-      def groups
-        (instrumentation['groups'] || {}).transform_keys do |key|
-          Regexp.new(Regexp.quote(key))
-        end
-      end
-
-      # Fetch the group name for a given scenario name. This will utilize the
-      # +SCENARIO_GROUPS+ map.
-      #
-      # @param name [String] the scenario name
-      # @return [String] the group name
-      def scenario_group(name)
-        groups.map do |pattern, group_name|
-          next unless pattern.match? name
-
-          group_name
-        end.compact.first || 'Various'
-      end
     end
   end
 end
